Add per-dcument preferences.
Introduce DocumentPreferences, whereby preferences are
first attempted to be read from the current document, before
falling back to the system preferences. This currently allows
one to specify the master document via modelines in the
tex file, e.g.
%
% gedit:master-document-filename = foo.tex
%
This also cleans up a lot of existing code

To unify this as a key-value store, I moved to ini style
files (no more XML). This is the first step towards cleaning
up the prefs a great deal more.

Future plans include introducing the concept of a
MasterDocumentPreferences, where, on a per-project/document basis,
one can customize the behaviour of the plugin, such as
 * Introducing more tools
 * Changing build actions
 * etc
tidy prefs, remove get_bool, defaults, etc